requirements.txt
matplotlib
pmdarima
prophet
psycopg2-binary
sqlalchemy
python-dotenv
pandas
wheel
scikit-learn
nbconvert
seaborn
traitlets
scipy
plotly
watchdog
mpld3
streamlit==1.12.0 # streamlit==1.11.1
Pillow
streamlit-autorefresh
jupyterflow
https://files.pythonhosted.org/packages/ab/88/dc597fdadd74b1a2fe7d4e4fb56f44a2b444a01b76ec35f31c7eee28f2d0/ray-2.1.0-cp310-cp310-manylinux2014_x86_64.whl #ray[default]==2.1.0
https://github.com/agapebondservant/ml-metrics-accelerator/zipball/main
lz4
pyyaml
prometheus-client
protobuf # protobuf==3.20.0 # protobuf==3.19.5
fastapi
#modin[ray]
uvicorn[standard]
gunicorn
boto3
tensorflow
rsocket
# asyncio==3.4.3
python-snappy
mlflow==2.0.1
filelock==3.9.0
pyarrow<7.0.0
hickle
python-multipart